+++ src/main/java/com/cefet/sistema_carcerario_digital/services/PessoaService.java
@@ -9,17 +9,20 @@ import org.springframework.transaction.annotation.Transactional;
 import com.cefet.sistema_carcerario_digital.dto.PessoaRequestDTO;
 import com.cefet.sistema_carcerario_digital.dto.PessoaResponseDTO;
 import com.cefet.sistema_carcerario_digital.entities.Pessoa;
-import com.cefet.sistema_carcerario_digital.exceptions.DatabaseException;
+import com.cefet.sistema_carcerario_digital.exceptions.BusinessException;
 import com.cefet.sistema_carcerario_digital.exceptions.ResourceNotFoundException;
+import com.cefet.sistema_carcerario_digital.repositories.CondenacaoRepository;
 import com.cefet.sistema_carcerario_digital.repositories.PessoaRepository;
 
 @Service
 public class PessoaService {
 
     private final PessoaRepository repo;
+    private final CondenacaoRepository condenacaoRepo;
 
-    PessoaService(PessoaRepository pessoaRepository) {
+    PessoaService(PessoaRepository pessoaRepository, CondenacaoRepository condenacaoRepository) {
         this.repo = pessoaRepository;
+        this.condenacaoRepo = condenacaoRepository;
     }
 
     @Transactional(readOnly = true)
@@ -39,9 +42,10 @@ public class PessoaService {
     @Transactional
     public PessoaResponseDTO inserir(PessoaRequestDTO dto) {
         String cpf = normalizarCpf(dto.getCpf());
+        validarCpf(cpf);
 
         if (repo.existsByCpf(cpf)) {
-            throw new DatabaseException("CPF já cadastrado.");
+            throw new BusinessException("CPF já cadastrado.");
         }
 
         Pessoa entity = new Pessoa();
@@ -58,9 +62,10 @@ public class PessoaService {
                 .orElseThrow(() -> new ResourceNotFoundException("Pessoa não encontrada. Id: " + id));
 
         String cpf = normalizarCpf(dto.getCpf());
+        validarCpf(cpf);
 
         if (repo.existsByCpfAndIdNot(cpf, id)) {
-            throw new DatabaseException("CPF já cadastrado.");
+            throw new BusinessException("CPF já cadastrado.");
         }
 
         copiarDtoParaEntidade(dto, entity, cpf);
@@ -74,6 +79,11 @@ public class PessoaService {
         if (!repo.existsById(id)) {
             throw new ResourceNotFoundException("Pessoa não encontrada. Id: " + id);
         }
+
+        if (condenacaoRepo.existsByPessoaId(id)) {
+            throw new BusinessException("Não é possível excluir pessoa com condenações vinculadas.");
+        }
+
         repo.deleteById(id);
     }
 
@@ -82,8 +92,13 @@ public class PessoaService {
         entity.setCpf(cpf);
     }
 
-    // -------- se der BO com o CPF pode ser aqui --------
     private String normalizarCpf(String cpf) {
         return cpf.replaceAll("\\D", "");
+    }
+
+    private void validarCpf(String cpf) {
+        if (cpf.length() != 11) {
+            throw new BusinessException("CPF deve conter 11 dígitos.");
+        }
     }
 }
